create fake controller
